[PJRT:Distributed] Avoid accidental narrowing conversion in coordination service.
May fix Windows build (https://github.com/google/jax/issues/11076)

PiperOrigin-RevId: 455016068

+++ tensorflow/compiler/xla/pjrt/distributed/service.cc
@@ -52,7 +52,7 @@ EnableCoordinationService(
   server_def.set_task_index(0);
   auto job_def = server_def.mutable_cluster()->add_job();
   job_def->set_name(job_name);
-  for (size_t i = 0; i < options.num_nodes; ++i) {
+  for (int32_t i = 0; i < options.num_nodes; ++i) {
     job_def->mutable_tasks()->insert({i, "UNKNOWN_SERVER_ADDRESS"});
   }
